# inbounds optimizations for CircularDeque and CircularBuffer