Fills a few inbounds optimization gaps in ```CircularDeque``` and ```CircularBuffer``` implementations, adding missing propagation and eliminating redundant checks.

Benchmarked with
```Julia
using DataStructures
using BenchmarkTools
function pushpop(xs)
    @inbounds push!(xs, 1)
    @inbounds popfirst!(xs)
    @inbounds pushfirst!(xs, 1)
    @inbounds pop!(xs)
end
function pushpop_checked(xs)
    push!(xs, 1)
    popfirst!(xs)
    pushfirst!(xs, 1)
    pop!(xs)
end
@btime pushpop($(CircularDeque{Int}(10)))
@btime pushpop_checked($(CircularDeque{Int}(10)))
@btime pushpop($(CircularBuffer{Int}(10)))
@btime pushpop_checked($(CircularBuffer{Int}(10)))
```

| Test         | Container        | Before | After |
|-----------|-----------------|-------:|------:|
| pushpop | CircularDeque |  7.6ns | 2.3ns |
| pushpop_checked | CircularDeque |  7.7ns | 6.3ns |
| pushpop | CircularBuffer |  7.3ns | 3.5ns |
| pushpop_checked | CircularBuffer |  7.4ns | 6.6ns |

Similar to #586 and #588